Update discord.yml
Changed it to NOT use the master YAML file due to errors. So it will use a local YAML file that points it to the right place. Seems simple enough honestly. The only downside is if I change something (like I want more or less info) I need to change it in every Github repo I am currently and actively using (up to 8 at any given time).